+++ src/BuildingBlocks/Application/InvalidCommandException.cs
@@ -10,7 +10,7 @@ namespace CompanyName.MyMeetings.BuildingBlocks.Application
         /// <summary>
         /// Gets the list of errors associated with the invalid command.
         /// </summary>
-        public ReadOnlyCollection<string> Errors { get; }
+        public List<string> Errors { get; }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="InvalidCommandException"/> class with the specified list of errors.
@@ -18,7 +18,7 @@ namespace CompanyName.MyMeetings.BuildingBlocks.Application
         /// <param name="errors">The list of errors associated with the invalid command.</param>
         public InvalidCommandException(List<string> errors)
         {
-            this.Errors = errors.AsReadOnly();
+            this.Errors = errors;
         }
     }
 }
