+++ src/Directory.Packages.props
@@ -19,8 +19,6 @@
   <ItemGroup Condition="$(MSBuildProjectName.EndsWith('BuildingBlocks.Application'))">
     <PackageReference Include="Dapper" Version="2.1.24" />
     <PackageReference Include="FluentValidation" Version="11.8.1" />
-    <PackageReference Include="IdentityServer4" Version="4.1.2" />
-    <PackageReference Include="IdentityServer4.AccessTokenValidation" Version="3.0.1" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
     <PackageReference Include="Quartz" Version="3.8.0" />
     <PackageReference Include="Serilog.AspNetCore" Version="8.0.0" />
