+++ src/Modules/UserAccess/Infrastructure/IdentityServer/ProfileService.cs
@@ -2,7 +2,7 @@
 using IdentityServer4.Models;
 using IdentityServer4.Services;
 
-namespace CompanyName.MyMeetings.Modules.UserAccess.Application.IdentityServer
+namespace CompanyName.MyMeetings.Modules.UserAccess.Infrastructure.IdentityServer
 {
     public class ProfileService : IProfileService
     {
